Update fromTop.1.0.js
Performance improvements: === doesn't do type casting, so it's faster than ==
One if() statement is cleaner than 2 if() statements(lines 27 and 29) and works just the same, feel free to revert this one :P
The ||(i.e. OR) operator at line 12 is cleaner, too. Instead of declaring the variable and then re-declaring it, it does it only once.